requirements/testing.txt
-r base.txt

black==23.7.0

pytest
pytest-django
pytest-cov
factory_boy
vcrpy==5.0.0
pytest-freezegun
pytest-mock
pytest-subtests
tomlkit
freezegun==1.2.2
django-debug-toolbar==4.1.0
curlylint==0.13.1
ruff==0.0.280
pytest-ruff
pytest-black
